requirements.txt
faiss-cpu==1.8.0.post1
numpy>=1.24
pandas>=2.0
sentence-transformers>=2.2.2
streamlit>=1.33.0
